+++ packages/component/src/Composer.js
@@ -3,150 +3,14 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 import Context from './Context';
-import createDeferred from './createDeferred';
-import retry from './retry';
-import sleep from './sleep';
-import spinWaitUntil from './spinWaitUntil';
-import timeout from './timeout';
-
-function createUtterance(utteranceLike, ponyfill) {
-  const { SpeechSynthesisUtterance } = ponyfill;
-  const {
-    lang,
-    onBoundary,
-    onEnd,
-    onError,
-    onStart,
-    pitch = 1,
-    rate = 1,
-    text,
-    voice,
-    volume = 1
-  } = utteranceLike;
-  const utterance = new SpeechSynthesisUtterance(text);
-  let targetVoice;
-
-  if (typeof voice === 'function') {
-    targetVoice = voice.call(speechSynthesis, getSerializableVoices(speechSynthesis));
-  } else {
-    const { voiceURI } = voice || {};
-
-    targetVoice = voiceURI && [].find.call([].slice.call(getSerializableVoices(speechSynthesis)), v => v.voiceURI === voiceURI);
-  }
-
-  // Edge will mute if "lang" is set to ""
-  utterance.lang = lang || '';
-
-  if (utterance.pitch || utterance.pitch === 0) {
-    utterance.pitch = pitch;
-  }
-
-  if (utterance.rate || utterance.rate === 0) {
-    utterance.rate = rate;
-  }
-
-  // Cognitive Services will error when "voice" is set to "null"
-  // Edge will error when "voice" is set to "undefined"
-  if (targetVoice) {
-    utterance.voice = targetVoice;
-  }
-
-  if (utterance.volume || utterance.volume === 0) {
-    utterance.volume = volume;
-  }
-
-  if (utterance.addEventListener) {
-    utterance.addEventListener('boundary', onBoundary);
-    utterance.addEventListener('end', onEnd);
-    utterance.addEventListener('error', onError);
-    utterance.addEventListener('start', onStart);
-  }
-
-  return utterance;
-}
-
-async function speakUtterance({ speechSynthesis }, { reject, resolve, utterance }) {
-  try {
-    const startDeferred = createDeferred();
-    const errorDeferred = createDeferred();
-    const endDeferred = createDeferred();
-
-    utterance.addEventListener('end', endDeferred.resolve);
-    utterance.addEventListener('error', errorDeferred.resolve);
-    utterance.addEventListener('start', startDeferred.resolve);
-
-    // if (speechSynthesis.speaking) {
-    //   console.warn(`ASSERTION: speechSynthesis.speaking should not be truthy before we call speak`);
-    // }
-
-    // Chrome quirks:
-    // 1. Speak an utterance
-    // 2. Cancel in the midway
-    // 3. Speak another utterance
-    // Expected: speaking is falsy, then turn to truthy, then receive "start" event, and audio played
-    // Actual: speaking is falsy, then turn to truthy (which is wrong), but receive no "start" event, and no audio played
-    // Workaround: retry 2 times with a second
-
-    // Safari quirks:
-    // - Audio doesn't play if the speech is started from a user event
-    // - If no audio is played, the "start" event won't fire
-
-    // For Chrome quirks, we need a custom queue, because we need to definitely know when to expect a "start" event.
-    // If we don't have a queue, the "start" event could be happening long time later because it's still pending in the queue.
-
-    // But with the custom queue, the first item might be started from non-user event. That means in Safari, the first item is muted.
-    // And after the first fail, the custom queue will play the second item from a non-user event code path. That means, all subsequent
-    // items are blocked until Safari has the very first item queued from user event.
-
-    // console.debug(`STARTING: ${ utterance.text }`);
-
-    await retry(async () => {
-      speechSynthesis.speak(utterance);
-
-      try {
-        await Promise.race([
-          startDeferred.promise,
-          timeout(1000)
-        ]);
-      } catch (error) {
-        // This is required for Chrome quirks.
-        // Chrome doesn't know it can't start speech, and it just wait there forever.
-        // We need to cancel it out.
-        speechSynthesis.cancel();
-
-        throw error;
-      }
-    }, 2, 0);
-
-    // console.debug(`STARTED: ${ utterance.text }`);
-
-    const endEvent = await Promise.race([
-      errorDeferred.promise,
-      endDeferred.promise,
-      spinWaitUntil(() => !speechSynthesis.speaking).then(() => sleep(500)).then(() => ({ type: 'end', artificial: true }))
-    ]);
-
-    // if (speechSynthesis.speaking) {
-    //   console.warn(`ASSERTION: speechSynthesis.speaking should not be truthy after speak is stopped`);
-    // }
-
-    // console.debug(`ENDED: ${ utterance.text }`);
-
-    if (endEvent.type === 'error') {
-      throw endEvent.error;
-    }
-
-    return resolve();
-  } catch (error) {
-    return reject(error);
-  }
-}
+import Utterance from './Utterance';
 
 class SpeechContext {
   constructor(ponyfill) {
     this.queueWithCurrent = [];
 
     this.cancel = this.cancel.bind(this);
+    this.cancelAll = this.cancelAll.bind(this);
     this.speak = this.speak.bind(this);
 
     this.setPonyfill(ponyfill);
@@ -156,92 +20,72 @@ class SpeechContext {
     this.ponyfill = { speechSynthesis, SpeechSynthesisUtterance };
   }
 
-  async cancel() {
+  async cancel(id) {
+    const index = this.queueWithCurrent.findIndex(utterance => utterance.id === id);
+
+    if (~index) {
+      return this.queueWithCurrent[index].cancel();
+    }
+  }
+
+  async cancelAll() {
     // console.debug(`CANCELLING QUEUED ITEMS: ${ this.queueWithCurrent.length }`);
 
-    this.queueWithCurrent.forEach(entry => entry.cancelled = true);
+    // this.queueWithCurrent.forEach(entry => entry.cancelled = true);
 
-    const cancelAll = Promise.all(this.queueWithCurrent.map(({ deferred: { promise } }) => promise.catch(err => 0)));
+    // const cancelAll = Promise.all(this.queueWithCurrent.map(({ deferred: { promise } }) => promise.catch(err => 0)));
 
-    this.ponyfill.speechSynthesis.cancel();
+    // this.ponyfill.speechSynthesis.cancel();
 
-    try {
-      await cancelAll;
-    } catch (err) {}
+    // try {
+    //   await cancelAll;
+    // } catch (err) {}
 
     // console.debug(`ALL CANCELLED OR FINISHED`);
   }
 
   speak(utteranceLike) {
-    const deferred = createDeferred();
-
     // console.debug(`QUEUED: ${ utteranceLike.text }`);
 
     if (
-      utteranceLike.uniqueID
-      && this.queueWithCurrent.find(({ utteranceLike: { uniqueID } }) => utteranceLike.uniqueID === uniqueID)
+      utteranceLike.id
+      && this.queueWithCurrent.find(({ id }) => id === utteranceLike.id)
     ) {
       // Do not queue duplicated speak with same unique ID
+      // console.debug('NOT QUEUEING DUPE');
+
       return;
     }
 
-    this.queueWithCurrent.push({
-      deferred,
-      utteranceLike
-    });
+    const utterance = new Utterance(utteranceLike);
+
+    this.queueWithCurrent = [...this.queueWithCurrent, utterance];
 
     if (this.queueWithCurrent.length === 1) {
       this._next();
     }
 
-    return deferred.promise;
+    return utterance.deferred.promise;
   }
 
   _next() {
-    const entry = this.queueWithCurrent[0];
+    const utterance = this.queueWithCurrent[0];
 
-    if (!entry) { return; }
+    if (!utterance) { return; }
 
-    entry.deferred.promise.then(() => {
-      this.queueWithCurrent.shift();
+    const { id } = utterance;
+    const promise = utterance.speak(this.ponyfill);
+
+    promise.then(() => {
+      this.queueWithCurrent = this.queueWithCurrent.filter(utterance => utterance.id !== id);
       this._next();
     }, () => {
       // TODO: If the error is due to Safari restriction on user touch
       //       The next loop on the next audio will also fail because it was not queued with a user touch
-      this.queueWithCurrent.shift();
+      this.queueWithCurrent = this.queueWithCurrent.filter(utterance => utterance.id !== id);
       this._next();
     });
-
-    if (entry.cancelled) {
-      // console.debug(`CANCELLED BEFORE PLAY: ${ entry.utteranceLike.text }`);
-
-      return entry.deferred.reject(new Error('cancelled'));
-    }
-
-    const utterance = createUtterance(entry.utteranceLike, this.ponyfill);
-
-    speakUtterance(this.ponyfill, {
-      reject: entry.deferred.reject,
-      resolve: entry.deferred.resolve,
-      utterance
-    });
   }
-}
-
-function getSerializableVoices(speechSynthesis) {
-  return speechSynthesis.getVoices().map(({
-    'default': def,
-    lang,
-    localService,
-    name,
-    voiceURI
-  }) => ({
-    'default': def,
-    lang,
-    localService,
-    name,
-    voiceURI
-  }));
 }
 
 export default class Composer extends React.Component {
@@ -263,7 +107,7 @@ export default class Composer extends React.Component {
         speechSynthesis: props.speechSynthesis,
         SpeechSynthesisUtterance: props.speechSynthesisUtterance
       }),
-      voices: getSerializableVoices(props.speechSynthesis)
+      voices: props.speechSynthesis.getVoices()
     };
   }
 
@@ -288,7 +132,7 @@ export default class Composer extends React.Component {
         nextProps.speechSynthesis.addEventListener && nextProps.speechSynthesis.addEventListener('voiceschanged', this.handleVoicesChanged);
       }
 
-      this.setState(() => ({ voices: getSerializableVoices(nextProps.speechSynthesis) }));
+      this.setState(() => ({ voices: nextProps.speechSynthesis.getVoices() }));
     }
   }
 
@@ -299,7 +143,7 @@ export default class Composer extends React.Component {
   }
 
   handleVoicesChanged({ target }) {
-    this.setState(() => ({ voices: getSerializableVoices(target) }));
+    this.setState(() => ({ voices: target.getVoices() }));
   }
 
   render() {