+++ packages/component/src/Say.js
@@ -10,7 +10,6 @@ const Say = props =>
     speechSynthesisUtterance={ props.speechSynthesisUtterance }
   >
     <SayPrimitive
-      exclusive={ props.exclusive }
       lang={ props.lang }
       onBoundary={ props.onBoundary }
       onEnd={ props.onEnd }
@@ -32,7 +31,6 @@ Say.defaultProps = {
 };
 
 Say.propTypes = {
-  exclusive: PropTypes.bool,
   lang: PropTypes.string,
   onBoundary: PropTypes.func,
   onEnd: PropTypes.func,