+++ packages/component/src/SayButton.js
@@ -1,41 +1,78 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import Context from './Context';
+import Composer from './Composer';
 
-const SayButton = props =>
-  <Context.Consumer
+class SayButton extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.handleClick = this.handleClick.bind(this);
+
+    this.state = {
+      busy: false,
+      id: Date.now() + Math.random()
+    };
+  }
+
+  componentWillUnmount() {
+    this.props.context.cancel(this.state.id);
+  }
+
+  handleClick() {
+    const { props, state } = this;
+
+    props.onClick && props.onClick(event);
+
+    this.setState(() => ({ busy: true }), async () => {
+      await props.context.speak({
+        id: state.id,
+        lang: props.lang,
+        onBoundary: props.onBoundary,
+        onEnd: props.onEnd,
+        onError: props.onError,
+        onStart: props.onStart,
+        pitch: props.pitch,
+        rate: props.rate,
+        text: props.speak,
+        voice: props.voice,
+        volume: props.volume
+      });
+
+      this.setState(() => ({
+        busy: false
+      }));
+    });
+  }
+
+  render() {
+    const {
+      props: { children, disabled },
+      state: { busy }
+    } = this;
+
+    return (
+      <button
+        disabled={ typeof disabled === 'boolean' ? disabled : busy }
+        onClick={ this.handleClick }
+      >
+        { children }
+      </button>
+    );
+  }
+}
+
+const SayButtonWithContext = props =>
+  <Composer
     speechSynthesis={ props.speechSynthesis }
     speechSynthesisUtterance={ props.speechSynthesisUtterance }
   >
     { context =>
-      <button onClick={ async event => {
-        if (props.exclusive) {
-          await context.cancel();
-        }
-
-        context.speak({
-          lang: props.lang,
-          onBoundary: props.onBoundary,
-          onEnd: props.onEnd,
-          onError: props.onError,
-          onStart: props.onStart,
-          pitch: props.pitch,
-          rate: props.rate,
-          text: props.speak,
-          voice: props.voice,
-          volume: props.volume
-        });
-
-        props.onClick && props.onClick(event);
-      } }>
-        { props.children }
-      </button>
+      <SayButton context={ context } { ...props } />
     }
-  </Context.Consumer>
+  </Composer>
 
-SayButton.propTypes = {
-  exclusive: PropTypes.bool,
+SayButtonWithContext.propTypes = {
   lang: PropTypes.string,
   onBoundary: PropTypes.func,
   onEnd: PropTypes.func,
@@ -48,4 +85,4 @@ SayButton.propTypes = {
   volume: PropTypes.number
 };
 
-export default SayButton;
+export default SayButtonWithContext