+++ packages/component/src/getSerializableVoices.js
@@ -1,0 +1,16 @@
+export default function (speechSynthesis) {
+  return speechSynthesis.getVoices();
+  // return speechSynthesis.getVoices().map(({
+  //   'default': def,
+  //   lang,
+  //   localService,
+  //   name,
+  //   voiceURI
+  // }) => ({
+  //   'default': def,
+  //   lang,
+  //   localService,
+  //   name,
+  //   voiceURI
+  // }));
+}