+++ packages/playground/src/App.js
@@ -1,39 +1,55 @@
+import { css } from 'glamor';
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
 
 import Say, { Composer, SayButton } from 'component';
 
-export default class App extends React.Component {
+const ROOT_CSS = css({
+  display: 'flex',
+
+  '& > section': {
+    flex: 1
+  }
+});
+
+const SEGMENTS = [
+  'A quick brown fox',
+  'jumped over',
+  'the lazy dogs',
+  'A quick brown fox jumped over the lazy dogs.'
+];
+
+export default class extends React.Component {
   constructor(props) {
     super(props);
 
-    this.handleAutoSpeakChange = this.handleAutoSpeakChange.bind(this);
-    this.handleExclusiveChange = this.handleExclusiveChange.bind(this);
-    this.handleVoiceChange = this.handleVoiceChange.bind(this);
+    this.handleAddClick = this.handleAddClick.bind(this);
+    this.handleRemoveFromQueue = this.handleRemoveFromQueue.bind(this);
+    this.handleSayEnd = this.handleRemoveFromQueue.bind(this);
+    this.handleSelectedVoiceChange = this.handleSelectedVoiceChange.bind(this);
 
     this.state = {
-      autoSpeak: null,
-      exclusive: false,
-      selectedVoice: null
+      queued: [],
+      selectedVoiceURI: null
     };
   }
 
-  handleAutoSpeakChange(nextAutoSpeak) {
-    this.setState(() => ({
-      autoSpeak: nextAutoSpeak
+  handleAddClick(text, pitch, rate) {
+    const id = Date.now() + Math.random();
+
+    this.setState(({ queued }) => ({
+      queued: [...queued, { id, pitch, rate, text }]
     }));
   }
 
-  handleExclusiveChange({ target: { checked } }) {
-    this.setState(() => ({
-      exclusive: checked
+  handleRemoveFromQueue(targetID) {
+    this.setState(({ queued }) => ({
+      queued: queued.filter(({ id }) => id !== targetID)
     }));
   }
 
-  handleVoiceChange(nextSelectedVoice, { target: { checked } }) {
-    checked && this.setState(() => ({
-      selectedVoice: nextSelectedVoice
+  handleSelectedVoiceChange(nextSelectedVoiceURI) {
+    this.setState(() => ({
+      selectedVoiceURI: nextSelectedVoiceURI
     }));
   }
 
@@ -42,130 +58,110 @@ export default class App extends React.Component {
 
     return (
       <Composer>
-        <div className="App">
-          <header className="App-header">
-            <img src={logo} className="App-logo" alt="logo" />
-            <h1 className="App-title">Welcome to React</h1>
-          </header>
-          <p className="App-intro">
-            To get started, edit <code>src/App.js</code> and save to reload.
-          </p>
-          <a href="https://github.com/compulim/react-say">
-            <img
-              alt="Fork me on GitHub"
-              src="https://s3.amazonaws.com/github/ribbons/forkme_right_gray_6d6d6d.png"
-              style={{ position: 'absolute', top: 0, right: 0, border: 0 }}
-            />
-          </a>
-          <Say speak="I am ready to say whatever you want me to." />
-          <Composer>
-            { ({ voices }) =>
-              <React.Fragment>
-                <h1>Options</h1>
-                <label>
-                  <input
-                    checked={ state.exclusive }
-                    onChange={ this.handleExclusiveChange }
-                    type="checkbox"
-                  />
-                  Exclusive
-                </label>
-                <h1>&lt;Say&gt;</h1>
-                {
-                  state.autoSpeak && <Say exclusive={ state.exclusive } speak={ state.autoSpeak } />
-                }
+        { ({ voices }) =>
+          <div className={ ROOT_CSS }>
+            <section className="words">
+              <article>
+                <header>
+                  <h1>Words</h1>
+                </header>
                 <ul>
-                  {
-                    ['A quick brown fox', 'jumped over', 'the lazy dogs'].map(text =>
-                      <li key={ text }>
-                        <label>
-                          <input
-                            checked={ state.autoSpeak === text }
-                            name="autosay"
-                            onChange={ this.handleAutoSpeakChange.bind(null, text) }
-                            type="radio"
-                          />
-                          { text }
-                        </label>
-                      </li>
-                    )
-                  }
+                  { SEGMENTS.map(segment =>
+                    <li key={ segment }>
+                      <button onClick={ this.handleAddClick.bind(null, segment, 1, 1) }><span role="img" aria-label="say">💬</span></button>
+                      <button onClick={ this.handleAddClick.bind(null, segment, 2, 1) }><span role="img" aria-label="high pitch">📈</span></button>
+                      <button onClick={ this.handleAddClick.bind(null, segment, .5, 1) }><span role="img" aria-label="low pitch">📉</span></button>
+                      <button onClick={ this.handleAddClick.bind(null, segment, 1, 2) }><span role="img" aria-label="fast">🐇</span></button>
+                      <button onClick={ this.handleAddClick.bind(null, segment, 1, .5) }><span role="img" aria-label="slow">🐢</span></button>
+                      &nbsp;{ segment }
+                    </li>
+                  ) }
                 </ul>
-                <h1>&lt;SayButton&gt;</h1>
-                <SayButton
-                  exclusive={ state.exclusive }
-                  speak="A quick brown fox jumped over the lazy dogs."
-                  voice={ state.selectedVoice }
-                >
-                  A quick brown fox jumped over the lazy dogs.
-                </SayButton>
-                <SayButton
-                  exclusive={ state.exclusive }
-                  speak="你想我噏乜呀？"
-                  voice={ voices => voices.find(voice => voice.lang === 'zh-HK') }
-                >
-                  你想我噏乜呀？
-                </SayButton>
-                <SayButton
-                  exclusive={ state.exclusive }
-                  pitch={ 2 }
-                  speak="High pitch"
-                  voice={ state.selectedVoice }
-                >
-                  High pitch
-                </SayButton>
-                <SayButton
-                  exclusive={ state.exclusive }
-                  pitch={ .5 }
-                  speak="Low pitch"
-                  voice={ state.selectedVoice }
-                >
-                  Low pitch
-                </SayButton>
-                <SayButton
-                  exclusive={ state.exclusive }
-                  speak="Low volume"
-                  voice={ state.selectedVoice }
-                  volume={ .1 }
-                >
-                  Low volume
-                </SayButton>
-                <SayButton
-                  exclusive={ state.exclusive }
-                  rate={ 2 }
-                  speak="High rate"
-                  voice={ state.selectedVoice }
-                >
-                  High rate
-                </SayButton>
-                <SayButton
-                  exclusive={ state.exclusive }
-                  rate={ .5 }
-                  speak="Low rate"
-                  voice={ state.selectedVoice }
-                >
-                  Low rate
-                </SayButton>
-                <h1>Voices</h1>
+              </article>
+              <article>
+                <header>
+                  <h1>Say immediate</h1>
+                </header>
                 <ul>
-                  { voices.map(voice =>
-                    <li key={ voice.voiceURI }>
+                  { SEGMENTS.map(segment =>
+                    <li key={ segment }>
+                      <SayButton speak={ segment }>{ segment }</SayButton>
+                    </li>
+                  ) }
+                </ul>
+              </article>
+              <article>
+                <header>
+                  <h1>Other languages</h1>
+                </header>
+                <ul>
+                  <li>
+                    <SayButton
+                      speak="一於記住一於記住每天向前望"
+                      voice={ voices => voices.find(voice => voice.lang === 'zh-HK') }
+                    >
+                      一於記住一於記住每天向前望
+                    </SayButton>
+                  </li>
+                  <li>
+                    <SayButton
+                      speak="お誕生日おめでとう"
+                      voice={ voices => voices.find(voice => voice.lang === 'ja-JP') }
+                    >
+                      お誕生日おめでとう
+                    </SayButton>
+                  </li>
+                </ul>
+              </article>
+            </section>
+            <section className="queue">
+              <article>
+                <header>
+                  <h1>Queue</h1>
+                </header>
+                { state.queued.length ?
+                    <ul>
+                      { state.queued.map(({ id, pitch, rate, text }) =>
+                        <li key={ id }>
+                          <button onClick={ this.handleRemoveFromQueue.bind(null, id) }>&times;</button>&nbsp;
+                          <span>{ text }</span>
+                          <Say
+                            onEnd={ this.handleSayEnd.bind(null, id) }
+                            pitch={ pitch }
+                            rate={ rate }
+                            speak={ text }
+                            voice={ voices => voices.find(({ voiceURI }) => voiceURI === state.selectedVoiceURI) }
+                          />
+                        </li>
+                      ) }
+                    </ul>
+                  :
+                    <div>Nothing queued</div>
+                }
+              </article>
+              <article>
+                <header>
+                  <h1>Available voices</h1>
+                </header>
+                <ul>
+                  { voices.map(({ lang, name, voiceURI }) =>
+                    <li key={ voiceURI }>
                       <label>
                         <input
-                          checked={ !!state.selectedVoice && voice.voiceURI === state.selectedVoice.voiceURI }
-                          onChange={ this.handleVoiceChange.bind(null, voice) }
-                          type="checkbox"
+                          checked={ voiceURI === state.selectedVoiceURI }
+                          onChange={ this.handleSelectedVoiceChange.bind(null, voiceURI) }
+                          name="voice"
+                          type="radio"
                         />
-                        ({ voice.lang }) { voice.name }
+                        [{ lang }] { name }
                       </label>
-                    </li>)
-                  }
+                    </li>
+                  ) }
                 </ul>
-                <pre>{ JSON.stringify(voices, null, 2) }</pre>
-              </React.Fragment>
-            }
-          </Composer>
-        </div>
+              </article>
+            </section>
+          </div>
+        }
       </Composer>
     );
   }